code changes on home fragment and home base fragment naming
